SocialNet/requirements.txt
Django==2.1
django-cors-middleware==1.3.1
django-role-permissions==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2==2.7
PyJWT==1.6.4
requests==2.20.1
factory-boy==2.11.1
email-hunter-python==1.1.0
